controller: get rid of v1 and extract releasecycle
Towards https://github.com/giantswarm/giantswarm/issues/5273
Fixes https://github.com/giantswarm/giantswarm/issues/5340

This is in prep to create release controller.
