+++ service/controller/releasecycle/resource/app/delete.go
