+++ service/controller/releasecycle/resource/app/desired.go
@@ -8,7 +8,7 @@ import (
 	"github.com/giantswarm/microerror"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/giantswarm/release-operator/service/controller/v1/key"
+	"github.com/giantswarm/release-operator/service/controller/releasecycle/key"
 )
 
 func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) (interface{}, error) {
