+++ service/controller/releasecycle/resource/app/resource.go
@@ -10,7 +10,7 @@ import (
 
 const (
 	// Name is the identifier of the resource.
-	Name = "appv1"
+	Name = "app"
 )
 
 // Config represents the configuration used to create a new app resource.
