+++ service/controller/releasecycle/resource/configmap/delete.go
